traceeval: Remove free_results()
The static function free_results() was never used. I kept it around for
when I was going to add it (and to remind myself not to forgot about it,
which is why I left the warning). But now that the code is being
rewritten, there's no more need to keep he warning around to remind me to
attach it. Get rid of it, all that code is going away anyway.

Link: https://lore.kernel.org/linux-trace-devel/20230804095219.41d66349@gandalf.local.home

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>