histograms: Add traceeval query
traceeval_query() fetches the values (aka results) of an entry with
keys that match the array of keys provided. The results of a query can
be useful for aggregating data between entries/trace events.

traceeval_results_release() frees the results fetched by the query. The
user must free the results of a successful query.

Link: https://lore.kernel.org/linux-trace-devel/20230809175340.3066-5-stevie.6strings@gmail.com

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Stevie Alvarez (Google) <stevie.6strings@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>