histograms: Add traceeval insert
traceeval_insert() updates the provided struct traceeval's histogram.
If an entry that exists with a keys field that match the keys argument,
the entries vals field are updated with a copy of the vals argument.
If such an entry does not exist, a new entry is added to the histogram,
with respect to the keys and vals arguments.

Link: https://lore.kernel.org/linux-trace-devel/20230809175340.3066-6-stevie.6strings@gmail.com

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Stevie Alvarez (Google) <stevie.6strings@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>