libtraceeval: Rename copy_traceeval_data_set() to dup_traceeval_data()
The name "copy_traceveal_data_set()" is misleading. It's actually
duplicating the data. The difference is that it allocates a new data
element and does not "copy" it. This is considered a "dup" and not a copy.
This makes it more in line of "strdup()".

It also removes the association of the order of the parameters, as when
something is being allocated, it usually comes at the end.

Link: https://lore.kernel.org/linux-trace-devel/20230817204528.114577-4-rostedt@goodmis.org

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>