libtraceeval histograms: Add traceeval struct to compare function
When looking at how this code would be implemented, I found that I needed
access to the traceeval structure within the compare callbacks. Pass that
in too.

Also, rearrange the parameters a little. The traceeval and type should go
first as they describe the "object", and the data should go last as they are
the values of the function.

Link: https://lore.kernel.org/linux-trace-devel/20230817204528.114577-8-rostedt@goodmis.org

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>