libtraceeval histogram: Have traceeval_query() just give the pointer to results
Instead of wasting allocation to return the results for queries,
especially since the query is likely not to modify the results, just give
a direct pointer to the entry->vals.

Mark it as const so (hopefully) nobody modifies it.

We still require the traceeval_results_release() to be called on it, as we
may in the future add a ref count to it and perhaps do a copy-on-write if
an update happens while something still has a hold on it.

Link: https://lore.kernel.org/linux-trace-devel/20230817204528.114577-19-rostedt@goodmis.org

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>