libtraceeval: Add size checks to insert and query functions
Currently, there's nothing that checks the size of the keys/values being
passed into traceeval_query() and traceeval_insert(). Make functions that
take the size of those arrays:

   traceeval_query_size()
   traceeval_insert_size()

and convert the traceeval_query() and traceeval_insert() into macros that
calculate the size of the fields to pass in:

  #define TRACEEVAL_ARRAY_SIZE(data)  \
     	(sizeof(data) / sizeof((data)[0]))

  #define traceeval_query(teval, keys, results) \
     traceeval_query_size(teval, TRACEEVAL_ARRAY_SIZE(keys), results)

and this will allow the code to check to make sure the size matches what
is expected (note, that currently is not done, but can be added).

If the keys or vals is not an array, but instead a pointer, if one of the
macros is used, the compiler will complain with:

 warning: division sizeof (const struct traceeval_data *) / sizeof (const struct traceeval_data)
          does not compute the number of array elements [-Wsizeof-pointer-div]
   20 | #define TRACEEVAL_ARRAY_SIZE(data)      (sizeof(data) / sizeof((data)[0]))
      |                                                       ^
 include/traceeval-hist.h:206:43: note: in expansion of macro TRACEEVAL_ARRAY_SIZE
  206 |         traceeval_query_size(teval, keys, TRACEEVAL_ARRAY_SIZE(keys), results)
      |                                           ^~~~~~~~~~~~~~~~~~~~

Then the user would need to call the "_size()" function directly
(as is done in the sample code, where the keys received by an iterator is
 used for a query)

Now that there are size functions for traceeval_insert() and
traceveal_query(), add checks to make sure that the size being passed in
is actually the size expected.

Link: https://lore.kernel.org/linux-trace-devel/20230927123314.989589-9-rostedt@goodmis.org

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>