libtraceeval: Add traceeval_init_data_size()
In the future, the traceeval_type and traceeval_data may need to include
more information and be expanded. If that happens, it will break backward
compatibility for older applications using the newer library.

To be able to handle backward compatibility in this situation, have the
traceeval_init() turn into traceeval_init_data_size() which takes the
sizeof struct traceeval_type and struct traceeval_data. Make
traceeval_init() into a macro that calls this function passing in the
sizeof(struct traceeval_type) and sizeof(struct traceeval_data). This way
if the sizes change for either one, the new code will know if the
application is using the new interface or the older one, and can
can keep the older one still functioning properly.

Link: https://lore.kernel.org/linux-trace-devel/20230927090904.396273fa@rorschach.local.home

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>