libtraceeval: Fix traceeval_insert() macro for NULL vals
If the vals is NULL (which is valid), the traceeval_insert() macro calls
traceeval_insert_size() with sizeof(vals)/sizeof(vals[0]) to get the
number of vals in the array. But this does not make sense with a NULL
value. Check for NULL and if vals is NULL then pass in zero.

Link: https://lore.kernel.org/linux-trace-devel/20230929051304.7b157905@rorschach.local.home

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>