libtraceeval: Move traceeval-hist.h to traceeval.h
Now that the API of the new histogram logic of libtraceeval is rather
stable, we can now remove the old version and replace it with the new
version. The traceeval-hist.h was just a tempory header until it was to
a point that it should be the default. It now is at that point. Remove
the old code of the histogram and use the new code.

Link: https://lore.kernel.org/linux-trace-devel/20230929082240.21420a2a@rorschach.local.home

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>