libtraceeval: Fix comparing unsigned against zero
nr_key_types and nr_val_types are both size_t which is an unsized number.
A compare against zero is always false:

	if (teval->nr_key_types < 0)

Change them to ssize_t (signed type).

Also change the few places the "size_t i" is used where it does:

	for (; i >= 0; i--)

as that too needs to be signed to work.

Link: https://lore.kernel.org/linux-trace-devel/20231003091032.06e697a3@gandalf.local.home

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>