libtraceeval: Remove need to use TRACEEVAL_TYPE_NONE in keys and vals
Currently, the size of the keys and vals are determined by the
requirement that the traceeval_type array end with an empty element of
type TRACEEVAL_TYPE_NONE.

To make the API be able to handle changes of the size of that structure,
the traceeval_init() was converted to a macro to pass in the sizeof of
that structure. This also means it can calculate the size of the array
with the TRACEEVAL_ARRAY_SIZE() macro.

Remove the need to have the initialization of traceeval keys and vals
have to add the NONE element at the end. It can still do so and this
will work just the same (the size will be determined by either the size
passed in or the first NONE element, which ever is smaller).

Link: https://lore.kernel.org/linux-trace-devel/20231003140802.1139616-2-rostedt@goodmis.org

Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Reviewed-by: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>