libtraceeval: Add wake-lat sample code
Add code that reads a trace.dat file and searches for the sched_waking and
sched_switch events. It then keeps track of the wake up latency of each
task as well as keeps statistics on that latency. At the end, it will
report the max/min total and count of each latency. For the max and min,
it also shows the timestamp of when they happened.

Link: https://lore.kernel.org/linux-trace-devel/20231006150851.480bd35e@gandalf.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>