libtraceeval: Have TRACEEVAL_SET_*() functions fail on wrong type
I've hit a few bugs using libtraceeval because of a common cut and paste
error. To assign keys, the following is common to do:

  TRACEEVAL_SET_STRING(keys[0], "string");
  TRACEEVAL_SET_NUMBER(keys[0], 123);

But the above has a bug! The second entry was supposed to be keys[1], and
not keys[0]. This bug had me scratching my head a few times until I found
it. The failure is that passing keys to traceeval_insert() or something
similar, the traceeval_insert() will fail because it's expecting keys[0]
to be a string, but the above converted it to a number.

Help prevent this by passing a string into one of the NUMBER() macros or a
number in the STRING() macro fail to build. It doesn't prevent all
combinations as it is common to pass various size to NUMBER() for example:

  unsigned long long val;
  TRACEEVAL_SET_NUMBER_16(keys[0], val);

The above is allowed, but a string should never go into a number and vice
versa. Catching that should find some bugs at build.

Nothing is done with TRACEEVAL_SET_POINTER() as pointers are totally the
responsibility of the developer.

This also requires that you can not just pass a void pointer to one of the
STRING() macros and it requires a typecast now (as the samples had to be
updated).

Link: https://lore.kernel.org/linux-trace-devel/20231007090740.7b0963cf@rorschach.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>