libtraceeval: Add new TRACEEVAL_TYPE_DELTA
Add a new traceeval_type called "delta". This type is a tuple of a delta
and a timestamp. It is a common occurrence to record a delta between
timestamps of two events and to record the delta with a stat. Instead of
creating two traceeval_data types (a number_64 timestamp and a number_64
delta), create a type that does that.

The TRACEEVAL_TYPE_DELTA is a structure of:

 struct {
	unsigned long long	delta;
	unsigned long long	timestamp;
 };

It automatically gets set as a STAT to retrieve statistics of the delta.
There's no need to supply another TIMESTAMP value as this will use the
given timestamp for the max and min saving.

Link: https://lore.kernel.org/linux-trace-devel/20231007122842.31152c71@rorschach.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>