libtraceeval: Add traceeval_iterator_delta_start_get()
Add APIs:

  traceeval_iterator_delta_start_get()
  traceeval_iterator_delta_stop()

After analysing events, there may be dangling deltas. That is, there may
be elements in a traceeval_delta that had a traceeval_delta_start() but no
matching traceeval_delta_stop(). Create an iterator that allows the
developer to find all these dangling elements.

 iter = traceeval_iterator_delta_start_get(teval);

Will return an iterator that will iterate over all the elements in the
delta that was started by not stopped.

 while (traceeval_iterator_next(iter, &keys) > 0) {

   traceeval_iterator_delta(stop, &results, ts, &delta, &start_ts);

Will iterator over all the dangling events, where the
traceeval_iterator_next() will get the keys of the dangling event and
traceeval_iterator_delta_stop() will retrieve the values, and the
"start_ts" that hold the timestamp of the traceeval_delta_start() and
passing in the last timestamp will also calculate the "delta".

Link: https://lore.kernel.org/linux-trace-devel/20231009025354.1577934-3-rostedt@goodmis.org

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>