libtraceeval: Do not just copy vals for delta elements
The traceeval_delta_start() on update just copied the values that were
passed in, but these elements are part of the internal traceeval. That
means on release, the strings will be freed. If the vals are simply
copied, then what was passed in is going to be freed.

Instead, do the same thing that is done on adding a new element, and call
traceeval_insert() to handle the update.

Link: https://lore.kernel.org/linux-trace-devel/20231010230246.74894849@rorschach.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>