libtraceeval: Fix timestamp when vals is NULL
If traceeval_init() is passed NULL for vals, it can crash because he
teval->timestamp_idx is not updated and stays at zero. Then the check for
timestamp will iterate the NULL array of vals and crash.

Move the initialization of teval->timestamp_idx right after the teval is
allocated and do not depend on the val check to update it.

Link: https://lore.kernel.org/linux-trace-devel/20231011182351.098416bb@gandalf.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>