libtraceeval: Simplify and correct man page example code
I pulled out each of the man page examples and compiled them as well as
ran them under valgrind and such. I restructured them to be a bit simpler
as well.

Also, I removed code examples that use STAT and TIMESTAMP flags. I'm
thinking of removing them (at least reserving them) and keeping them out
of the API before release. That's because the TRACEEVAL_TYPE_DELTA appears
to be all that is needed (for now).

Note, these flags are still explained. If I do remove them, the commit that
removes them will then remove the comments about them in the man pages.

Link: https://lore.kernel.org/linux-trace-devel/20231011223814.1fada6a1@gandalf.local.home

Cc: Ross Zwisler <zwisler@google.com>
Cc: Stevie Alvarez <stevie.6strings@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>