libtraceeval: Remove TIMESTAMP and STAT flags from API
The TRACEEVAL_TYPE_DELTA can be used to keep track of numbers that require
maximum, minimum, count, total, average and stand deviation. This is
easier to use than to create a separate TRACEEVAL_FL_TIMESTAMP to keep
track of the time associated with a value that has a STAT flag.

Note, the logic is only removed from the API, that is, the traceeval_flags
no longer have them. They are now replied with:

  TRACEEVAL_FL_RESERVE_3
  TRACEEVAL_FL_RESERVE_4

The logic in the code still keeps it around in case we discover that
there's use cases where the TRACEEVAL_TYPE_DELTA does not suffice. But we
also have the option of simply removing it too.

The man pages have been updated to remove any mention of these flags, and
to use TRACEEVAL_TYPE_DELTA.

Also, added man pages for:

  traceeval_stat_average()
  traceeval_stat_stddev()

Link: https://lore.kernel.org/linux-trace-devel/20231012105801.3ca2c789@gandalf.local.home

Cc: Ross Zwisler <zwisler@google.com>
Cc: Stevie Alvarez   <stevie.6strings@gmail.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>