libtraceeval task-eval: Simplify logic by adding more helper functions
Instead of trying to keep track of all the traceeval_data array sizes in
the code. Add wrappers that assign each of the various traceevals.

 insert_cpu_data()
 start_cpu_data()
 continue_cpu_data()
 stop_cpu_data()
 cpu_last_state()

 insert_task_data()
 start_task_data()
 stop_task_data()
 task_data_exists()

 insert_thread_data()

The above functions handle the traceveal logic with the keys and the vals
and only the necessary data is passed to them. This makes it much easier
to handle the logic in the main code.

Link: https://lore.kernel.org/linux-trace-devel/20231012124016.26ea05fb@gandalf.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>