libtraceeval: Remove sizeof() checks from traceeval_data_create_size()
The first version of the traceeval_delta required the user to create the
type externally. That was cumbersome and was changed to simply use:

  traceeval_delta_create()

to create a delta element of an existing traceeval. As the original
version did not depend on an existing traceeval, it had to have the
sizeof checks on the traceeval structures in case they ever change in the
future.

Now that the code requires an existing traceeval to create a
traceeval_delta, the checks have already been done at the creation of the
traceeval itself. No need to repeat the checks in traceeval_data_create().

Link: https://lore.kernel.org/linux-trace-devel/20231012141845.68116b67@gandalf.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>