libtraceeval: Add timestamp to traceeval_delta_query()
When writing the man pages for traceeval_delta_query(), I realized that it
should also return the timestamp associated to the delta. This may be
useful as well, and it should be added before the API is set.

The timestamp parameter may be NULL if it is to be ignored.

If it is retrieved, it will be the last value of a traceeval_delta_stop()
or traceeval_delta_continue() if that was the last function called on the
give keys. If traceeval_delat_stop() is called last, then the timestamp
will be zero.

Link: https://lore.kernel.org/linux-trace-devel/20231012145139.4b66f9cf@gandalf.local.home

Cc: Ross Zwisler <zwisler@google.com>
Signed-off-by: Steven Rostedt (Google) <rostedt@goodmis.org>